Tl pacijent linkovi #1 3

+++ Backend/src/main/java/isaps/tim18/PSW_ISA_ClinicalC_2019/service/PosetaService.java
@@ -22,7 +22,7 @@ public class PosetaService {
         return posetaRepo.findAll();
     }
 
-    public List<Poseta> findByPacijent_id(int id){
+    public List<Poseta> findByPacijent_id(Long id){
         return posetaRepo.findByPacijent_id(id);
     }
